refactor(zfb-types): collapse parallel Segment enums into one canonical type
Create zfb-types crate with the canonical Segment enum (Static / Dynamic
/ Catchall), replacing the two parallel definitions that existed in:

- crates/zfb-router/src/route.rs  (the original, with Hash + Serde)
- crates/zfb-render/src/paths.rs  (the "temporary local stub" noted in
  its doc comment)

Both crates now re-export zfb_types::Segment at their original paths
(zfb_router::Segment and zfb_render::paths::Segment) so downstream
callers are unaffected. The zfb-types Segment preserves all four derives
from the router original: Debug, Clone, PartialEq, Eq, Hash, Serialize,
Deserialize.
