+++ crates/zfb-types/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "zfb-types"
+version = "0.0.0"
+edition = "2021"
+license = "MIT"
+description = "zfb-types: canonical shared data types used across multiple zfb crates"
+
+[dependencies]
+serde = { workspace = true }
